[❗️Fix] RefreshManager 테스트 완료
