+++ Pill-IT/Presentation/View/MainTabBarController.swift
@@ -17,6 +17,7 @@ final class MainTabBarController: WHTabbarController {
     private var secondVC : PillNotificationViewController!
     private let userNotificationCenter = UNUserNotificationCenter.current()
     private let repository = RealmRepository()
+    private let refresh = RefreshManager.shared
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -32,7 +33,7 @@ final class MainTabBarController: WHTabbarController {
             selectedIndex = 0
         }
         
-        timerForResetNotification()
+        refresh.timerForResetNotification()
     }
     
     override func viewDidLayoutSubviews() {
@@ -57,28 +58,6 @@ final class MainTabBarController: WHTabbarController {
             present(nav, animated: true)
         }
     }
-    
-    @objc private func resetNotificationAction() {
-        print("✅ 타이머 실행")
-        
-        let todayDate = Date()
-        let yesterDate = Calendar.current.date(byAdding: .day, value: -1, to: Date())!
-
-        if let todayPillAlarmDateTable = repository.fetchPillAlarmDateItemIsDone(alaramDate: todayDate) {
-            // 현재 날짜의 모든 알림 등록
-            userNotificationCenter.addNotificationRequest(byList: todayPillAlarmDateTable)
-            userNotificationCenter.printPendingNotification()
-        } else { print("오늘의 알림이 없습니다 ✅") }
-        
-        if let yesterDatePillAlarmDateTable = repository.fetchPillAlarmDateItemIsDone(alaramDate: yesterDate) {
-            // 어제 날짜의 모든 알림 삭제
-            userNotificationCenter.removeAllNotification(by: yesterDatePillAlarmDateTable)
-        } else { print("어제의 알림이 없습니다 ✅") }
-
-        // 목록 확인
-        userNotificationCenter.printPendingNotification()
-    }
-    
 }
 
 //MARK: - Timer 설정하기
@@ -92,20 +71,6 @@ extension MainTabBarController {
         secondVC = PillNotificationViewController()
         
         setViewControllers([firstNav, secondVC], animated: true) // tab view 설정
-    }
-    
-    func timerForResetNotification() {
-        let calendar = Calendar.current
-        let now = Date()
-        let date = calendar.date(
-            bySettingHour: 00,
-            minute: 00,
-            second: 00,
-            of: now)!
-        
-        let timer = Timer(fireAt: date, interval: 0, target: self, selector: #selector(resetNotificationAction), userInfo: nil, repeats: false)
-        
-        RunLoop.main.add(timer, forMode: RunLoop.Mode.common)
     }
 }
